+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.5.3
 
+* `Versions`:
+    * `Kotlin`: `1.5.0` -> `1.5.10`
 * `Coroutines`:
     * Extensions `doInUI` and `doInDefault` were replaced in common and available on any supported platform
     * Extension `doInIO` replaced into `jvm` and available on any `JVM` platform